requirements.txt
coverage==7.0.5
Flask==2.2.2
pydantic==1.10.4
pytest==7.2.1
python-dotenv==0.21.1
SQLAlchemy==1.4.46
typing_extensions==4.4.0